requirements.txt
google-api-python-client==2.94.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
selenium==4.15.2